# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Regulate AI - Reporting Bridge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/company-summary:
    get:
      operationId: getCompanySummary
      tags: [dashboard]
      summary: Company dashboard summary
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySummary"

  # ── Reports ────────────────────────────────────────────────────────────────
...
  /reports/{reportId}/schema:
    get:
      operationId: listReportColumns
      tags: [reportSchema]
      summary: List all schema columns for a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of columns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportColumn"
    post:
      operationId: createReportColumn
      tags: [reportSchema]
      summary: Add a column to a report schema
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportColumnInput"
      responses:
        "201":
          description: Created column
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportColumn"
  /reports/{reportId}/schema/{columnId}:
    delete:
      operationId: deleteReportColumn
      tags: [reportSchema]
      summary: Remove a column from a report schema
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
        - name: columnId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Report Rules ───────────────────────────────────────────────────────────
  /reports/{reportId}/rules:
    get:
      operationId: listReportRules
      tags: [reportRules]
      summary: List validation rules for a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportRule"
    post:
      operationId: createReportRule
      tags: [reportRules]
      summary: Add a validation rule to a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRuleInput"
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRule"
  /reports/{reportId}/rules/{ruleId}:
    delete:
      operationId: deleteReportRule
      tags: [reportRules]
      summary: Remove a validation rule
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
        - name: ruleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Companies ──────────────────────────────────────────────────────────────
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all registered companies
      parameters:
        - name: domain
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Register a new financial company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # ── Report Requests (Company Side) ────────────────────────────────────────
...
components:
  schemas:
...
    CompanySummary:
      type: object
      properties:
        companyName:
          type: string
        pendingAcceptance:
          type: integer
        accepted:
          type: integer
        validating:
          type: integer
        submitted:
          type: integer
        overdue:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - companyName
        - pendingAcceptance
        - accepted
        - validating
        - submitted
        - overdue
        - recentActivity
    ActivityItem:
      type: object
      properties:
        requestId:
          type: integer
        reportName:
          type: string
        status:
          type: string
        updatedAt:
          type: string
      required:
        - requestId
        - reportName
        - status
        - updatedAt
...
    ReportColumn:
      type: object
      properties:
        id:
          type: integer
        reportId:
          type: integer
        columnName:
          type: string
        dataType:
          type: string
          enum: [string, integer, decimal, date, boolean, enum]
        description:
          type: string
        isMandatory:
          type: boolean
        enumValues:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - reportId
        - columnName
        - dataType
        - description
        - isMandatory
        - createdAt
    ReportColumnInput:
      type: object
      properties:
        columnName:
          type: string
          minLength: 1
        dataType:
          type: string
          enum: [string, integer, decimal, date, boolean, enum]
        description:
          type: string
        isMandatory:
          type: boolean
        enumValues:
          type: string
      required:
        - columnName
        - dataType
        - description
        - isMandatory

    # ── Report Rules ──────────────────────────────────────────────────────────
    ReportRule:
      type: object
      properties:
        id:
          type: integer
        reportId:
          type: integer
        columnName:
          type: string
        ruleType:
          type: string
          enum: [range, not_null, regex, cross_column, enum_check]
        operator:
          type: ["string", "null"]
        value:
          type: ["string", "null"]
        compareColumn:
          type: ["string", "null"]
        description:
          type: string
        createdAt:
          type: string
      required:
        - id
        - reportId
        - columnName
        - ruleType
        - description
        - createdAt
    ReportRuleInput:
      type: object
      properties:
        columnName:
          type: string
          minLength: 1
        ruleType:
          type: string
          enum: [range, not_null, regex, cross_column, enum_check]
        operator:
          type: string
        value:
          type: string
        compareColumn:
          type: string
        description:
          type: string
      required:
        - columnName
        - ruleType
        - description

    # ── Company ────────────────────────────────────────────────────────────────
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: string
          enum: [brokerage, asset_management, financing, leasing, factoring, fintech_bnpl, other]
        email:
          type: string
        contactPerson:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - domain
        - email
        - contactPerson
        - createdAt
    CompanyInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        domain:
          type: string
          enum: [brokerage, asset_management, financing, leasing, factoring, fintech_bnpl, other]
        email:
          type: string
        contactPerson:
          type: string
      required:
        - name
        - domain
        - email
        - contactPerson

    # ── Report Request ─────────────────────────────────────────────────────────
...