# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Regulate AI - Reporting Bridge API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  requestId: number;
  reportName: string;
  status: string;
  updatedAt: string;
}