# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Regulate AI - Reporting Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyDomain } from './companyDomain';

export interface Company {
  id: number;
  name: string;
  domain: CompanyDomain;
  email: string;
  contactPerson: string;
  createdAt: string;
}