# lib/api-zod/src/generated/types/reportColumn.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Regulate AI - Reporting Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportColumnDataType } from './reportColumnDataType';

export interface ReportColumn {
  id: number;
  reportId: number;
  columnName: string;
  dataType: ReportColumnDataType;
  description: string;
  isMandatory: boolean;
  /** @nullable */
  enumValues?: string | null;
  createdAt: string;
}