# lib/api-zod/src/generated/types/reportColumnInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Regulate AI - Reporting Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportColumnInputDataType } from './reportColumnInputDataType';

export interface ReportColumnInput {
  /** @minLength 1 */
  columnName: string;
  dataType: ReportColumnInputDataType;
  description: string;
  isMandatory: boolean;
  enumValues?: string;
}